grouped dropdown preference overloads
