+++ miuix-preference/src/commonMain/kotlin/top/yukonga/miuix/kmp/preference/OverlayDropdownPreference.kt
@@ -26,7 +26,9 @@ import top.yukonga.miuix.kmp.basic.BasicComponentDefaults
 import top.yukonga.miuix.kmp.basic.DropdownArrowEndAction
 import top.yukonga.miuix.kmp.basic.DropdownColors
 import top.yukonga.miuix.kmp.basic.DropdownDefaults
+import top.yukonga.miuix.kmp.basic.DropdownEntry
 import top.yukonga.miuix.kmp.basic.DropdownImpl
+import top.yukonga.miuix.kmp.basic.HorizontalDivider
 import top.yukonga.miuix.kmp.basic.ListPopupColumn
 import top.yukonga.miuix.kmp.basic.PopupPositionProvider
 import top.yukonga.miuix.kmp.basic.Text
@@ -73,8 +75,81 @@ fun OverlayDropdownPreference(
     enabled: Boolean = true,
     showValue: Boolean = true,
     renderInRootScaffold: Boolean = true,
+    // collapseOnSelection: Boolean = true,
     onExpandedChange: ((Boolean) -> Unit)? = null,
     onSelectedIndexChange: ((Int) -> Unit)? = null,
+) {
+    val entry = remember(
+        items, selectedIndex, onSelectedIndexChange,
+    ) { DropdownEntry(items, selectedIndex, onSelectedIndexChange) }
+    return OverlayDropdownPreference(
+        entry = entry,
+        title = title,
+        modifier = modifier,
+        titleColor = titleColor,
+        summary = summary,
+        summaryColor = summaryColor,
+        dropdownColors = dropdownColors,
+        startAction = startAction,
+        bottomAction = bottomAction,
+        insideMargin = insideMargin,
+        maxHeight = maxHeight,
+        enabled = enabled,
+        showValue = showValue,
+        renderInRootScaffold = renderInRootScaffold,
+        collapseOnSelection = true,
+        onExpandedChange = onExpandedChange,
+    )
+}
+
+@Composable
+private fun OverlayDropdownPreferencePopup(
+    items: List<String>,
+    selectedIndex: Int,
+    isDropdownExpanded: Boolean,
+    onDismiss: () -> Unit,
+    onDismissFinished: () -> Unit,
+    maxHeight: Dp?,
+    dropdownColors: DropdownColors,
+    hapticFeedback: HapticFeedback,
+    renderInRootScaffold: Boolean,
+    // collapseOnSelection: Boolean,
+    onSelectedIndexChange: ((Int) -> Unit)?,
+) {
+    val entry = remember(
+        items, selectedIndex, onSelectedIndexChange,
+    ) { DropdownEntry(items, selectedIndex, onSelectedIndexChange) }
+    return OverlayDropdownPreferencePopup(
+        entry = entry,
+        isDropdownExpanded = isDropdownExpanded,
+        onDismiss = onDismiss,
+        onDismissFinished = onDismissFinished,
+        maxHeight = maxHeight,
+        dropdownColors = dropdownColors,
+        hapticFeedback = hapticFeedback,
+        renderInRootScaffold = renderInRootScaffold,
+        collapseOnSelection = true,
+    )
+}
+
+@Composable
+fun OverlayDropdownPreference(
+    entry: DropdownEntry,
+    title: String,
+    modifier: Modifier = Modifier,
+    titleColor: BasicComponentColors = BasicComponentDefaults.titleColor(),
+    summary: String? = null,
+    summaryColor: BasicComponentColors = BasicComponentDefaults.summaryColor(),
+    dropdownColors: DropdownColors = DropdownDefaults.dropdownColors(),
+    startAction: @Composable (() -> Unit)? = null,
+    bottomAction: (@Composable () -> Unit)? = null,
+    insideMargin: PaddingValues = BasicComponentDefaults.InsideMargin,
+    maxHeight: Dp? = null,
+    enabled: Boolean = true,
+    showValue: Boolean = true,
+    renderInRootScaffold: Boolean = true,
+    collapseOnSelection: Boolean = true,
+    onExpandedChange: ((Boolean) -> Unit)? = null,
 ) {
     val interactionSource = remember { MutableInteractionSource() }
     val isDropdownExpanded = remember { mutableStateOf(false) }
@@ -91,7 +166,7 @@ fun OverlayDropdownPreference(
         }
     }
 
-    val itemsNotEmpty = items.isNotEmpty()
+    val itemsNotEmpty = entry.items.isNotEmpty()
     val actualEnabled = enabled && itemsNotEmpty
 
     val actionColor = if (actualEnabled) {
@@ -123,24 +198,24 @@ fun OverlayDropdownPreference(
         startAction = startAction,
         endActions = {
             if (showValue && itemsNotEmpty) {
-                Text(
-                    text = items[selectedIndex],
-                    modifier = Modifier
-                        .padding(end = 8.dp)
-                        .align(Alignment.CenterVertically)
-                        .weight(1f, fill = false),
-                    fontSize = MiuixTheme.textStyles.body2.fontSize,
-                    color = actionColor,
-                    textAlign = TextAlign.End,
-                )
+                if (entry.selectedIndex != null)
+                    Text(
+                        text = entry.items[entry.selectedIndex!!],
+                        modifier = Modifier
+                            .padding(end = 8.dp)
+                            .align(Alignment.CenterVertically)
+                            .weight(1f, fill = false),
+                        fontSize = MiuixTheme.textStyles.body2.fontSize,
+                        color = actionColor,
+                        textAlign = TextAlign.End,
+                    )
             }
             DropdownArrowEndAction(
                 actionColor = actionColor,
             )
             if (itemsNotEmpty) {
                 OverlayDropdownPreferencePopup(
-                    items = items,
-                    selectedIndex = selectedIndex,
+                    entry = entry,
                     isDropdownExpanded = isDropdownExpanded.value,
                     onDismiss = { setExpanded(false) },
                     onDismissFinished = { isHoldDown.value = false },
@@ -148,7 +223,7 @@ fun OverlayDropdownPreference(
                     dropdownColors = dropdownColors,
                     hapticFeedback = hapticFeedback,
                     renderInRootScaffold = renderInRootScaffold,
-                    onSelectedIndexChange = onSelectedIndexChange,
+                    collapseOnSelection = collapseOnSelection,
                 )
             }
         },
@@ -161,8 +236,7 @@ fun OverlayDropdownPreference(
 
 @Composable
 private fun OverlayDropdownPreferencePopup(
-    items: List<String>,
-    selectedIndex: Int,
+    entry: DropdownEntry,
     isDropdownExpanded: Boolean,
     onDismiss: () -> Unit,
     onDismissFinished: () -> Unit,
@@ -170,16 +244,19 @@ private fun OverlayDropdownPreferencePopup(
     dropdownColors: DropdownColors,
     hapticFeedback: HapticFeedback,
     renderInRootScaffold: Boolean,
-    onSelectedIndexChange: ((Int) -> Unit)?,
+    collapseOnSelection: Boolean,
 ) {
-    val onSelectState = rememberUpdatedState(onSelectedIndexChange)
+    val currentEntry by rememberUpdatedState(entry)
+    val currentCollapseOnSelection by rememberUpdatedState(collapseOnSelection)
     val currentOnDismiss by rememberUpdatedState(onDismiss)
     val currentHapticFeedback by rememberUpdatedState(hapticFeedback)
     val onItemSelected: (Int) -> Unit = remember {
         { selectedIdx ->
             currentHapticFeedback.performHapticFeedback(HapticFeedbackType.Confirm)
-            onSelectState.value?.invoke(selectedIdx)
-            currentOnDismiss()
+            currentEntry.onSelectedIndexChange?.let { it(selectedIdx) }
+            if (currentCollapseOnSelection) {
+                currentOnDismiss()
+            }
         }
     }
     OverlayListPopup(
@@ -191,15 +268,175 @@ private fun OverlayDropdownPreferencePopup(
         renderInRootScaffold = renderInRootScaffold,
     ) {
         ListPopupColumn {
-            items.forEachIndexed { index, string ->
+            entry.items.forEachIndexed { index, string ->
                 key(index) {
                     DropdownImpl(
                         text = string,
-                        optionSize = items.size,
-                        isSelected = selectedIndex == index,
+                        optionSize = entry.items.size,
+                        isSelected = entry.selectedIndex == index,
                         dropdownColors = dropdownColors,
                         onSelectedIndexChange = onItemSelected,
                         index = index,
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun OverlayDropdownPreference(
+    entries: List<DropdownEntry>,
+    title: String,
+    modifier: Modifier = Modifier,
+    titleColor: BasicComponentColors = BasicComponentDefaults.titleColor(),
+    summary: String? = null,
+    summaryColor: BasicComponentColors = BasicComponentDefaults.summaryColor(),
+    dropdownColors: DropdownColors = DropdownDefaults.dropdownColors(),
+    startAction: @Composable (() -> Unit)? = null,
+    bottomAction: (@Composable () -> Unit)? = null,
+    insideMargin: PaddingValues = BasicComponentDefaults.InsideMargin,
+    maxHeight: Dp? = null,
+    enabled: Boolean = true,
+    showValue: Boolean = true,
+    renderInRootScaffold: Boolean = true,
+    collapseOnSelection: Boolean = false,
+    onExpandedChange: ((Boolean) -> Unit)? = null,
+) {
+    val interactionSource = remember { MutableInteractionSource() }
+    val isDropdownExpanded = remember { mutableStateOf(false) }
+    val isHoldDown = remember { mutableStateOf(false) }
+    val hapticFeedback = LocalHapticFeedback.current
+    val currentHapticFeedback by rememberUpdatedState(hapticFeedback)
+    val currentOnExpandedChange = rememberUpdatedState(onExpandedChange)
+    val setExpanded: (Boolean) -> Unit = remember {
+        { expanded ->
+            if (isDropdownExpanded.value != expanded) {
+                isDropdownExpanded.value = expanded
+                currentOnExpandedChange.value?.invoke(expanded)
+            }
+        }
+    }
+
+    val entriesNotEmpty = entries.isNotEmpty() && entries.any { it.items.isNotEmpty() }
+    val actualEnabled = enabled && entriesNotEmpty
+
+    val actionColor = if (actualEnabled) {
+        MiuixTheme.colorScheme.onSurfaceVariantActions
+    } else {
+        MiuixTheme.colorScheme.disabledOnSecondaryVariant
+    }
+
+    val handleClick = remember(actualEnabled) {
+        {
+            if (actualEnabled) {
+                setExpanded(!isDropdownExpanded.value)
+                if (isDropdownExpanded.value) {
+                    isHoldDown.value = true
+                    currentHapticFeedback.performHapticFeedback(HapticFeedbackType.ContextClick)
+                }
+            }
+        }
+    }
+
+    BasicComponent(
+        modifier = modifier,
+        interactionSource = interactionSource,
+        insideMargin = insideMargin,
+        title = title,
+        titleColor = titleColor,
+        summary = summary,
+        summaryColor = summaryColor,
+        startAction = startAction,
+        endActions = {
+            val selectedValueText = entries.joinToString("\n") { group ->
+                group.selectedIndex?.let { index -> group.items.getOrNull(index) } ?: ""
+            }.ifBlank { null }
+            if (showValue && entriesNotEmpty && !selectedValueText.isNullOrBlank()) {
+                Text(
+                    text = selectedValueText,
+                    modifier = Modifier.padding(end = 8.dp),
+                    fontSize = MiuixTheme.textStyles.body2.fontSize,
+                    color = actionColor,
+                    textAlign = TextAlign.End,
+                    lineHeight = MiuixTheme.textStyles.body2.lineHeight,
+                )
+            }
+            DropdownArrowEndAction(
+                actionColor = actionColor,
+            )
+            if (entriesNotEmpty) {
+                OverlayDropdownPreferencePopup(
+                    entries = entries,
+                    isDropdownExpanded = isDropdownExpanded.value,
+                    onDismiss = { setExpanded(false) },
+                    onDismissFinished = { isHoldDown.value = false },
+                    maxHeight = maxHeight,
+                    dropdownColors = dropdownColors,
+                    hapticFeedback = hapticFeedback,
+                    renderInRootScaffold = renderInRootScaffold,
+                    collapseOnSelection = collapseOnSelection,
+                )
+            }
+        },
+        bottomAction = bottomAction,
+        onClick = handleClick,
+        holdDownState = isHoldDown.value,
+        enabled = actualEnabled,
+    )
+}
+
+@Composable
+private fun OverlayDropdownPreferencePopup(
+    entries: List<DropdownEntry>,
+    isDropdownExpanded: Boolean,
+    onDismiss: () -> Unit,
+    onDismissFinished: () -> Unit,
+    maxHeight: Dp?,
+    dropdownColors: DropdownColors,
+    hapticFeedback: HapticFeedback,
+    renderInRootScaffold: Boolean,
+    collapseOnSelection: Boolean,
+) {
+    val currentEntries by rememberUpdatedState(entries)
+    val currentCollapseOnSelection by rememberUpdatedState(collapseOnSelection)
+    val currentOnDismiss by rememberUpdatedState(onDismiss)
+    val currentHapticFeedback by rememberUpdatedState(hapticFeedback)
+    val onItemSelected: (Int, Int) -> Unit = remember {
+        { entryIdx, selectedIdx ->
+            currentHapticFeedback.performHapticFeedback(HapticFeedbackType.Confirm)
+            currentEntries.getOrNull(entryIdx)?.onSelectedIndexChange?.invoke(selectedIdx)
+            if (currentCollapseOnSelection) {
+                currentOnDismiss()
+            }
+        }
+    }
+    OverlayListPopup(
+        show = isDropdownExpanded,
+        alignment = PopupPositionProvider.Align.End,
+        onDismissRequest = onDismiss,
+        onDismissFinished = onDismissFinished,
+        maxHeight = maxHeight,
+        renderInRootScaffold = renderInRootScaffold,
+    ) {
+        ListPopupColumn {
+            entries.forEachIndexed { entryIdx, entry ->
+                entry.items.forEachIndexed { itemIdx, option ->
+                    DropdownImpl(
+                        text = option,
+                        optionSize = entry.items.size,
+                        isSelected = entry.selectedIndex == itemIdx,
+                        dropdownColors = dropdownColors,
+                        onSelectedIndexChange = { selectedIdx ->
+                            onItemSelected(entryIdx, selectedIdx)
+                        },
+                        index = itemIdx,
+                    )
+                }
+                if (entryIdx != entries.lastIndex) {
+                    HorizontalDivider(
+                        modifier = Modifier.padding(horizontal = 20.dp),
+                        thickness = 1.dp,
                     )
                 }
             }
