--- miuix-preference/src/commonMain/kotlin/top/yukonga/miuix/kmp/preference/OverlayMultiGroupDropdownPreference.kt
@@ -1,194 +1,0 @@
-// Copyright 2026, compose-miuix-ui contributors
-// SPDX-License-Identifier: Apache-2.0
-
-package top.yukonga.miuix.kmp.preference
-
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.hapticfeedback.HapticFeedback
-import androidx.compose.ui.hapticfeedback.HapticFeedbackType
-import androidx.compose.ui.platform.LocalHapticFeedback
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-import top.yukonga.miuix.kmp.basic.BasicComponent
-import top.yukonga.miuix.kmp.basic.BasicComponentColors
-import top.yukonga.miuix.kmp.basic.BasicComponentDefaults
-import top.yukonga.miuix.kmp.basic.DropdownArrowEndAction
-import top.yukonga.miuix.kmp.basic.DropdownColors
-import top.yukonga.miuix.kmp.basic.DropdownDefaults
-import top.yukonga.miuix.kmp.basic.DropdownImpl
-import top.yukonga.miuix.kmp.basic.HorizontalDivider
-import top.yukonga.miuix.kmp.basic.ListPopupColumn
-import top.yukonga.miuix.kmp.basic.MultiGroupDropdownEntry
-import top.yukonga.miuix.kmp.basic.PopupPositionProvider
-import top.yukonga.miuix.kmp.basic.Text
-import top.yukonga.miuix.kmp.overlay.OverlayListPopup
-import top.yukonga.miuix.kmp.theme.MiuixTheme
-
-@Composable
-fun OverlayMultiGroupDropdownPreference(
-    entries: List<MultiGroupDropdownEntry>,
-    title: String,
-    modifier: Modifier = Modifier,
-    titleColor: BasicComponentColors = BasicComponentDefaults.titleColor(),
-    summary: String? = null,
-    summaryColor: BasicComponentColors = BasicComponentDefaults.summaryColor(),
-    dropdownColors: DropdownColors = DropdownDefaults.dropdownColors(),
-    startAction: @Composable (() -> Unit)? = null,
-    bottomAction: (@Composable () -> Unit)? = null,
-    insideMargin: PaddingValues = BasicComponentDefaults.InsideMargin,
-    maxHeight: Dp? = null,
-    enabled: Boolean = true,
-    showValue: Boolean = true,
-    renderInRootScaffold: Boolean = true,
-    collapseOnSelection: Boolean = false,
-    onExpandedChange: ((Boolean) -> Unit)? = null,
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    val isDropdownExpanded = remember { mutableStateOf(false) }
-    val isHoldDown = remember { mutableStateOf(false) }
-    val hapticFeedback = LocalHapticFeedback.current
-    val currentHapticFeedback by rememberUpdatedState(hapticFeedback)
-    val currentOnExpandedChange = rememberUpdatedState(onExpandedChange)
-    val setExpanded: (Boolean) -> Unit = remember {
-        { expanded ->
-            if (isDropdownExpanded.value != expanded) {
-                isDropdownExpanded.value = expanded
-                currentOnExpandedChange.value?.invoke(expanded)
-            }
-        }
-    }
-
-    val entriesNotEmpty = entries.isNotEmpty() && entries.any { it.items.isNotEmpty() }
-    val actualEnabled = enabled && entriesNotEmpty
-
-    val actionColor = if (actualEnabled) {
-        MiuixTheme.colorScheme.onSurfaceVariantActions
-    } else {
-        MiuixTheme.colorScheme.disabledOnSecondaryVariant
-    }
-
-    val handleClick = remember(actualEnabled) {
-        {
-            if (actualEnabled) {
-                setExpanded(!isDropdownExpanded.value)
-                if (isDropdownExpanded.value) {
-                    isHoldDown.value = true
-                    currentHapticFeedback.performHapticFeedback(HapticFeedbackType.ContextClick)
-                }
-            }
-        }
-    }
-
-    BasicComponent(
-        modifier = modifier,
-        interactionSource = interactionSource,
-        insideMargin = insideMargin,
-        title = title,
-        titleColor = titleColor,
-        summary = summary,
-        summaryColor = summaryColor,
-        startAction = startAction,
-        endActions = {
-            val selectedValueText = entries.joinToString("\n") { group ->
-                group.items.getOrNull(group.selectedIndex).orEmpty()
-            }.ifBlank { null }
-            if (showValue && entriesNotEmpty && !selectedValueText.isNullOrBlank()) {
-                Text(
-                    text = selectedValueText,
-                    modifier = Modifier.padding(end = 8.dp),
-                    fontSize = MiuixTheme.textStyles.body2.fontSize,
-                    color = actionColor,
-                    textAlign = TextAlign.End,
-                    lineHeight = MiuixTheme.textStyles.body2.lineHeight,
-                )
-            }
-            DropdownArrowEndAction(
-                actionColor = actionColor,
-            )
-            if (entriesNotEmpty) {
-                OverlayMultiGroupDropdownPreferencePopup(
-                    entries = entries,
-                    isDropdownExpanded = isDropdownExpanded.value,
-                    onDismiss = { setExpanded(false) },
-                    onDismissFinished = { isHoldDown.value = false },
-                    maxHeight = maxHeight,
-                    dropdownColors = dropdownColors,
-                    hapticFeedback = hapticFeedback,
-                    renderInRootScaffold = renderInRootScaffold,
-                    collapseOnSelection = collapseOnSelection,
-                )
-            }
-        },
-        bottomAction = bottomAction,
-        onClick = handleClick,
-        holdDownState = isHoldDown.value,
-        enabled = actualEnabled,
-    )
-}
-
-@Composable
-private fun OverlayMultiGroupDropdownPreferencePopup(
-    entries: List<MultiGroupDropdownEntry>,
-    isDropdownExpanded: Boolean,
-    onDismiss: () -> Unit,
-    onDismissFinished: () -> Unit,
-    maxHeight: Dp?,
-    dropdownColors: DropdownColors,
-    hapticFeedback: HapticFeedback,
-    renderInRootScaffold: Boolean,
-    collapseOnSelection: Boolean,
-) {
-    val currentEntries by rememberUpdatedState(entries)
-    val currentCollapseOnSelection by rememberUpdatedState(collapseOnSelection)
-    val currentOnDismiss by rememberUpdatedState(onDismiss)
-    val currentHapticFeedback by rememberUpdatedState(hapticFeedback)
-    val onItemSelected: (Int, Int) -> Unit = remember {
-        { entryIdx, selectedIdx ->
-            currentHapticFeedback.performHapticFeedback(HapticFeedbackType.Confirm)
-            currentEntries.getOrNull(entryIdx)?.onSelectedIndexChange?.invoke(selectedIdx)
-            if (currentCollapseOnSelection) {
-                currentOnDismiss()
-            }
-        }
-    }
-    OverlayListPopup(
-        show = isDropdownExpanded,
-        alignment = PopupPositionProvider.Align.End,
-        onDismissRequest = onDismiss,
-        onDismissFinished = onDismissFinished,
-        maxHeight = maxHeight,
-        renderInRootScaffold = renderInRootScaffold,
-    ) {
-        ListPopupColumn {
-            entries.forEachIndexed { entryIdx, entry ->
-                entry.items.forEachIndexed { itemIdx, option ->
-                    DropdownImpl(
-                        text = option,
-                        optionSize = entry.items.size,
-                        isSelected = entry.selectedIndex == itemIdx,
-                        dropdownColors = dropdownColors,
-                        onSelectedIndexChange = { selectedIdx ->
-                            onItemSelected(entryIdx, selectedIdx)
-                        },
-                        index = itemIdx,
-                    )
-                }
-                if (entryIdx != entries.lastIndex) {
-                    HorizontalDivider(
-                        modifier = Modifier.padding(horizontal = 20.dp),
-                        thickness = 1.dp,
-                    )
-                }
-            }
-        }
-    }
-}
